Move DependencyResolutionFactory to Factories namespace.
